requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
web3==6.12.0
aiohttp==3.9.1
asyncio==3.4.3
APScheduler==3.10.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
base58==2.1.1  
